Refactor routes into separate files
